dspy_agent/requirements.txt
dspy>=2.3.6
fastapi
python-dotenv
uvicorn
pydantic 